+++ smurfs/src/index.js
@@ -6,7 +6,7 @@ import thunk from 'redux-thunk';
 import logger from 'redux-logger';
 import { createStore, applyMiddleware } from 'redux';
 import { Provider } from 'react-redux';
-import rootReducer from './reducers';
+import  rootReducer  from './reducers';
 
 const store = createStore(rootReducer, applyMiddleware(thunk,logger));// this is the most basic reducer. A function that returns and object. Replace it.
